Change how pagination of posts works
- Use kaminari pagination
- Create a paginator in user_graph_type for user posts.
- Define a paginator and selector for public posts
- Got rid of time conversions except for milliseconds, made that the default

I tried to figure out how one could get at some of the other kaminari methods from the single posts resolving method, but I could not figure it out. I'll probably end up with a paginator
resolver for the counts, etc., that also returns an array of the paged objects.

The public_posts method joins and includes the user record for each public post to save on the N+1 queries.